Handle inflate returning Z_OK at end of stream
The inflate function may return Z_OK rather than Z_STREAM_END when the last
bytes of the stream are inflated. In this case inflate returns Z_STREAM_END on
the next call. However in this situation we will already have reset avail_out
for the next line and the "d_stream.avail_out == 0" condition will be false.